+++ examples/basic-pwa/src/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+const App = () => {
+  return <h2>Hello World</h2>;
+};
+
+ReactDOM.render(<App />, document.querySelector('#ice-container'));