+++ packages/plugin-app-core/CHANGELOG.md
@@ -9,6 +9,15 @@
 - [feat] refactor runtime folder structure
 - [feat] api of render template
 
+## 1.4.10
+
+- [fix] globby pattern do not support path of win32
+
+## 1.4.9
+
+- [fix] ignore project type when get source file
+- [feat] register method of getSourceFile
+
 ## 1.4.8
 
 - [feat] add onPageNotFound and onUnhandledRejection types in ejs