+++ packages/plugin-icestark/CHANGELOG.md
@@ -5,6 +5,15 @@
 - [feat] migrate runtime api
 - [feat] upgrade build-scripts
 
+## 2.2.1
+
+- [fix] move all identifiers to visitor to avoid wrong execution results.
+
+# 2.2.0
+
+- [feat] support setting custom lifecycles. ([#337](https://github.com/ice-lab/icestark/issues/337))
+- [fix] append lifecycles for other framework.
+
 # 2.1.0
 
 - [fix] get rendering container through `props` instead of `getMountNode`. ([#227](https://github.com/ice-lab/icestark/issues/227))